# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boleteria Online",
    "description": "Colleción de API's correspondientes al sistema de Pasajes Online.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@boleteriasystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuario/login": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuario/ban": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoginResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/LoginResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RegistroRequest": {
        "required": [
          "confirm_password",
          "dni",
          "email",
          "fecha_nacimiento",
          "first_name",
          "last_name",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          },
          "confirm_password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "dni": {
            "type": "integer",
            "format": "int64"
          },
          "fecha_nacimiento": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "UsuarioResponse": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "tipo": {
            "$ref": "#/components/schemas/UsuarioTipo"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsuarioResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/UsuarioResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsuarioTipo": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}